+++ test/post-deploy.test.mjs
@@ -10,8 +10,8 @@
  * governing permissions and limitations under the License.
  */
 /* eslint-env mocha */
-import { describe, it } from 'node:test';
 import assert from 'assert';
+import { describe, it } from 'node:test';
 
 [
   {
@@ -175,13 +175,21 @@ import assert from 'assert';
       if (!process.env.TEST_INTEGRATION) {
         this.skip();
       }
+
+      const startTime = Date.now();
       const response = await fetch(`https://${domain}/.rum/@adobe/helix-rum-js@^1/src/index.js`, {
         method: 'GET',
       });
       assert.strictEqual(response.status, 200);
+      assert(Date.now() - startTime < 1000, 'Response took too long');
+
       // eslint-disable-next-line no-unused-expressions
       assert.match(response.headers.get('content-type'), /^text\/javascript/);
       assert.strictEqual(response.headers.get('x-frame-options'), 'DENY');
+      assert(
+        response.headers.get('x-rum-trace').startsWith('be-'),
+        'This request cannot be served by the helix backend, so should use the package registries',
+      );
     });
 
     it('rum js module is served with compression', async function test() {
@@ -211,6 +219,7 @@ import assert from 'assert';
       const ccHeader = respRange.headers.get('cache-control').split(',');
       assert(ccHeader.find((header) => header.trim() === 'max-age=3600'), 'Should have a max cache age of 3600');
       assert.strictEqual(respRange.headers.get('x-frame-options'), 'DENY');
+      assert.strictEqual(respRange.headers.get('x-rum-trace'), 'hlx');
 
       const respSpecific = await fetch(`https://${domain}/.rum/@adobe/helix-rum-enhancer@2.33.0/src/index.js`);
       assert.strictEqual(respSpecific.status, 200);
@@ -219,6 +228,42 @@ import assert from 'assert';
       const maxAge = Number(maHeader.split('=')[1]);
       assert(maxAge > 3600, 'Should have a max cache age greater than 3600');
       assert.strictEqual(respSpecific.headers.get('x-frame-options'), 'DENY');
+    });
+
+    it('rum js is served from helix backend', async function test() {
+      if (!process.env.TEST_INTEGRATION) {
+        this.skip();
+      }
+
+      const startTime = Date.now();
+      const respRange = await fetch(`https://${domain}/.rum/@adobe/helix-rum-js@~2.11.4/dist/rum-standalone.js`);
+      assert.strictEqual(respRange.status, 200);
+      assert.strictEqual('hlx', respRange.headers.get('x-rum-trace'));
+      assert(Date.now() - startTime < 1000, 'Response took too long');
+
+      const startTime2 = Date.now();
+      const respSpecific = await fetch(`https://${domain}/.rum/@adobe/helix-rum-js@2.11.4/dist/rum-standalone.js`);
+      assert.strictEqual(respSpecific.status, 200);
+      assert.strictEqual('hlx', respSpecific.headers.get('x-rum-trace'));
+      assert(Date.now() - startTime2 < 1000, 'Response took too long');
+    });
+
+    it('rum enhancer is served from helix backend', async function test() {
+      if (!process.env.TEST_INTEGRATION) {
+        this.skip();
+      }
+
+      const startTime = Date.now();
+      const respRange = await fetch(`https://${domain}/.rum/@adobe/helix-rum-enhancer@%5E2/src/index.js`);
+      assert.strictEqual(respRange.status, 200);
+      assert.strictEqual('hlx', respRange.headers.get('x-rum-trace'));
+      assert(Date.now() - startTime < 1000, 'Response took too long');
+
+      const startTime2 = Date.now();
+      const respSpecific = await fetch(`https://${domain}/.rum/@adobe/helix-rum-enhancer@2.34.3/src/index.js`);
+      assert.strictEqual(respSpecific.status, 200);
+      assert.strictEqual('hlx', respSpecific.headers.get('x-rum-trace'));
+      assert(Date.now() - startTime2 < 1000, 'Response took too long');
     });
 
     it.skip('rum js module is being served with default replacements', async function test() {
@@ -234,6 +279,7 @@ import assert from 'assert';
       const text = await response.text();
       assert.include(text, 'adobe-helix-rum-js-1.0.0');
       assert.strictEqual(response.headers.get('x-frame-options'), 'DENY');
+      assert.strictEqual(response.headers.get('x-rum-trace'), 'hlx');
     });
 
     it('Missing id returns 400', async function test() {
@@ -348,6 +394,30 @@ import assert from 'assert';
       assert.strictEqual(resp.headers.get('x-frame-options'), 'DENY');
     });
 
+    it('Reject paths that contain partially encoded ". ."', async function test() {
+      if (!process.env.TEST_INTEGRATION) {
+        this.skip();
+      }
+
+      const resp = await fetch(`https://${domain}/.rum/web-vitals/.%09./web-vitalsxyz/demo.html?pkgreg=unpkg`);
+      assert.strictEqual(resp.status, 400);
+      const respTxt = await resp.text();
+      assert(respTxt.startsWith('Invalid path'));
+      assert.strictEqual(resp.headers.get('x-frame-options'), 'DENY');
+    });
+
+    it('Reject paths that contain partially encoded ". ." with additional characters', async function test() {
+      if (!process.env.TEST_INTEGRATION) {
+        this.skip();
+      }
+
+      const resp = await fetch(`https://${domain}/.rum/web-vitals/.%09.%2Fweb-vitalsxyz%2FDEMO.html%23%5E?pkgreg=unpkg`);
+      assert.strictEqual(resp.status, 400);
+      const respTxt = await resp.text();
+      assert(respTxt.startsWith('Invalid path'));
+      assert.strictEqual(resp.headers.get('x-frame-options'), 'DENY');
+    });
+
     it('Non-existent files in .rum directory return 404', async function test() {
       if (!process.env.TEST_INTEGRATION) {
         this.skip();
@@ -357,6 +427,38 @@ import assert from 'assert';
       const respTxt = await resp.text();
       assert.strictEqual(resp.status, 404, `Expected 404 but got ${resp.status}. Response body: ${respTxt}`);
       assert.strictEqual(resp.headers.get('x-frame-options'), 'DENY');
+    });
+
+    it('CORS headers are set for helix-rum-enhancer cwv plugin', async function test() {
+      if (!process.env.TEST_INTEGRATION) {
+        this.skip();
+      }
+      const response = await fetch(`https://${domain}/.rum/@adobe/helix-rum-enhancer@%5E2/src/plugins/cwv.js`, {
+        method: 'GET',
+      });
+      assert.strictEqual(response.status, 200);
+      assert.strictEqual(response.headers.get('access-control-allow-origin'), '*');
+      assert.strictEqual(response.headers.get('access-control-allow-methods'), 'GET, HEAD, OPTIONS');
+      assert.strictEqual(response.headers.get('access-control-allow-headers'), '*');
+      assert.strictEqual(response.headers.get('access-control-expose-headers'), '*');
+      assert.strictEqual(response.headers.get('x-frame-options'), 'DENY');
+      assert.strictEqual(response.headers.get('x-rum-trace'), 'hlx');
+    });
+
+    it('CORS headers are set for helix-rum-enhancer webcomponent plugin', async function test() {
+      if (!process.env.TEST_INTEGRATION) {
+        this.skip();
+      }
+      const response = await fetch(`https://${domain}/.rum/@adobe/helix-rum-enhancer@%5E2/src/plugins/webcomponent.js`, {
+        method: 'GET',
+      });
+      assert.strictEqual(response.status, 200);
+      assert.strictEqual(response.headers.get('access-control-allow-origin'), '*');
+      assert.strictEqual(response.headers.get('access-control-allow-methods'), 'GET, HEAD, OPTIONS');
+      assert.strictEqual(response.headers.get('access-control-allow-headers'), '*');
+      assert.strictEqual(response.headers.get('access-control-expose-headers'), '*');
+      assert.strictEqual(response.headers.get('x-frame-options'), 'DENY');
+      assert.strictEqual(response.headers.get('x-rum-trace'), 'hlx');
     });
   });
 });